% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenFIGI_Func.R
\name{OpenFIGI_assignadditionalcols}
\alias{OpenFIGI_assignadditionalcols}
\title{Assign addition cols}
\usage{
OpenFIGI_assignadditionalcols(rst1, additioncols)
}
\description{
Assign addition cols
}
\details{
internal function
}

