% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenFIGI_Func.R
\name{sampleOpenFIGIdf}
\alias{sampleOpenFIGIdf}
\title{Generate a Sample data.frame for FIGI request}
\usage{
sampleOpenFIGIdf()
}
\value{
a data.frame of sample request
}
\description{
Generate a Sample data.frame for FIGI request
}

