% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenFIGI_Func.R
\name{showFIGIIDType}
\alias{showFIGIIDType}
\title{Generate FIGI Supported Identifiers}
\usage{
showFIGIIDType()
}
\value{
a char vector
}
\description{
Generate FIGI Supported Identifiers
}

