% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenFIGI_Func.R
\name{OpenFIGI_assignadditionalcols}
\alias{OpenFIGI_assignadditionalcols}
\title{Assign addition cols}
\usage{
OpenFIGI_assignadditionalcols(rst1, additioncols)
}
\arguments{
\item{rst1}{a result list}

\item{additioncols}{additional columns you would like to include in the data.frame}
}
\value{
a data.frame with new columns
}
\description{
Assign addition cols
}
\details{
internal function
}

