% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROpenFIGI.R
\docType{package}
\name{ROpenFIGI}
\alias{ROpenFIGI}
\title{ROpenFIGI}
\details{
Provide a simple interface to OpenFIGI by Bloomberg.
parse data into data.frame
}
\author{
Ruokun Huang \email{hruokun.2008@gmail.com}

Maintainer: Ruokun Huang \email{hruokun.2008@gmail.com}
}
